spanner: convert to datastore v2 API
